feat: add MicrositePaymentController for handling transactions by microsite
- Added a new file `MicrositePaymentController.php` in the `app/Http/Controllers` directory.
- The `MicrositePaymentController` class handles the retrieval of transactions by microsite.
- It retrieves payments associated with a specific microsite ID and includes the microsite information.
- The controller uses the `Payment` and `Microsites` models.
- The `transactionsByMicrosite` method returns an Inertia response with the payments and microsites data.

Refactor PaymentController to remove transactionsByMicrosite method

- Removed the `transactionsByMicrosite` method from the `PaymentController` class.
- This method is now handled by the newly added `MicrositePaymentController`.

Remove unnecessary code in LoginRequest

- Removed unused imports and commented out code in the `LoginRequest` class.

Remove unnecessary validation rules in StorePaymentRequest

- Removed commented out validation rules in the `StorePaymentRequest` class.

Update PlacetoPayGateway to use new API endpoint

- Updated the `PlacetoPayGateway` class to use a new API endpoint for processing payments.
- The `process` method now sends a POST request to the updated endpoint.

Update PaymentService to remove unnecessary code

- Removed unnecessary code in the `PaymentService` class.
- The `create` method no longer includes an empty process.

Update routes to use MicrositePaymentController

- Updated the route for handling transactions by microsite to use the `MicrositePaymentController` instead of the `PaymentController`.

Update StorePaymentTest to assert correct status and errors

- Updated the `testItHandlesClientError` and `testItHandlesUnknownError` methods in the `StorePaymentTest` class.
- The tests now assert the correct status code and session errors.
