+++ app/Http/Controllers/PaymentController.php
@@ -29,7 +29,6 @@ class PaymentController extends Controller
             'payment' => $payment,
             'gateway' => $gateway,
         ]);
-
         $response = $paymentService->create($buyerData);
         if ($response->status === 'exception') {
             return back()->withErrors(['message' => $response->message]);
@@ -69,18 +68,4 @@ class PaymentController extends Controller
         ]);
     }
 
-    public function transactionsByMicrosite($microsite_id): \Inertia\Response
-    {
-        $user = Auth::user();
-        $payments = Payment::where('microsite_id', $microsite_id)
-            ->with('microsite')
-            ->get();
-
-        $microsites = Microsites::all();
-
-        return Inertia::render('Payments/Transactions', [
-            'payments' => $payments,
-            'microsites' => $microsites,
-        ]);
-    }
 }
