+++ app/Http/Requests/StorePaymentRequest.php
@@ -23,16 +23,6 @@ class StorePaymentRequest extends FormRequest
     public function rules(): array
     {
         return [
-            // 'description' => ['required', 'string', 'min:3', 'max:100'],
-            // 'amount' => ['required', 'integer', 'min:1', 'max:999999999999'],
-            // 'microsite_id' => ['required', 'numeric', 'exists:microsites,id'],
-            // 'currency' => ['required', Rule::in(Currency::toArray())],
-            // 'name' => ['required', 'string', 'min:3', 'max:20'],
-            // 'last_name' => ['required', 'string', 'min:3', 'max:20'],
-            // 'email' => ['required', 'string', 'email'],
-            // 'document_number' => ['required', 'numeric', 'digits_between:6,20'],
-            // 'document_type' => ['required', Rule::in(DocumentTypes::toArray())],
-            // 'gateway' => ['required', Rule::in(PaymentGateway::toArray())],
         ];
     }
 }
