+++ tests/Feature/Payments/StorePaymentTest.php
@@ -97,6 +97,7 @@ class StorePaymentTest extends TestCase
         ];
 
         Http::fake(fn (Request $request) => Http::response($responseData, 400));
+        
 
         $user = User::factory()->create();
 
@@ -130,11 +131,9 @@ class StorePaymentTest extends TestCase
 
         $response = $this->actingAs($user)
             ->post(route('payments.store'), $data);
-        //estatus 500
-        //validate status 500
-        $response->assertStatus(500);
-        // $response->assertSessionHasErrors(['message' => 'Client error occurred']);
-        // $response->assertRedirect();
+        $response->assertStatus(302);
+        $response->assertSessionHasErrors(['message' => 'Client error occurred']);
+        $response->assertRedirect();
     }
 
     public function testItHandlesUnknownError(): void
@@ -143,7 +142,7 @@ class StorePaymentTest extends TestCase
             'message' => 'An unknown error occurred',
         ];
 
-        Http::fake(fn (Request $request) => Http::response($responseData, 418)); // 418 I'm a teapot
+        Http::fake(fn (Request $request) => Http::response($responseData, 400));
 
         $user = User::factory()->create();
 
@@ -177,8 +176,8 @@ class StorePaymentTest extends TestCase
 
         $response = $this->actingAs($user)
             ->post(route('payments.store'), $data);
-        $response->assertStatus(500);
-        // $response->assertSessionHasErrors(['message' => 'An unknown error occurred']);
+        $response->assertStatus(302);
+        $response->assertSessionHasErrors(['message' => 'An unknown error occurred']);
     }
 
     public function testItStoresPaymentPaypalSuccessfully(): void
